add chooseRoute component
